Fix #8: Unexpected inability to use *-exchange-point-and-mark-motion
When moving to the bounds, the mark was not set if the cursor was
already at the bounds.

This meant it wasn't possible define a bounds then run
meep-exchange-point-and-mark-motion as a way to select the bounds
when the cursor happened to already be at the bounds.

Now moving to bounds always sets the mark, even it didn't move from the
previous point, making exchange-point-and-mark-motion work as expected.