Initial "meep-transpose" implementation
Add a transpose key "v" to the default keymap, allowing any motion
to be followed by a transpose request.

From my own usage the most common kinds of actions are transposing lines
and symbols, although transposing characters works too.

Note that this is somewhat work-in-progress, however the basic
functionality is useful enough that I think it's worth including as-is,
improvements can be made as part of general development.