New feature: implement character bounds region-marking
Add new functions:
- meep-region-mark-bounds-of-char-{inner, outer}
- meep-region-mark-bounds-of-char-contextual-{inner, outer}

This allows an equivalent of VIM's (change-in-quotes) `ci'`,
to be performed with the same number of key-strokes.

In this case the commands needed are ordered [mark-in, quote, change].
once the

- meep-region-mark-bounds-of-char-inner
- '
- meep-insert-change

A single quote is and changing are just examples, once there is region
any other operation can be performed.

The contextual version of this command searches common character pairs
and sets the bounds the the closest one.

For convenience, RET can be used for contextual marking,
[meep-region-mark-bounds-of-char-inner, RET]

Allowing `o-RET` (with the default keymap) to contextually mark in
the quotes or brackets.

The character pairs are customizable.

Note that proper bracket matching has not yet been implemented,
so properly matching brackets that include nesting is not yet
working as expected.

Resolves #13.